Rework inputs. File version: March7-10am
I’ve redesigned the Current Visit tabs. Split into two tabs because one was too large for the optimal processing in the VUE system.  It also helps to just have less to see.  Also redid the whole way we do keys.  Its a bit more manual but it’s a lot more straight forward.  Colour coded too.  Blue: page level.  Yellowish: table level.  Mauve: Fieldset level.  To put a field into a table it needs to have a level2Key AND a row number in tableColumn.  To put a field into a fieldset it needs a level4Key AND the row/col numbers in  fsetRow/fsetCol.

I hope this new approach is easier to understand and work with.  I’m happy that this approach gets rids of a ton of fragile formulas that seemed to always get corrupted when we added and removed rows.  Now a spreadsheet column contains either text or formula but not both. And if it has a formula the entire column has the same formula.