Stacked tables
Based on the significant refactoring of the page definitions.  TableCells are now focused on just the table display but now we can also create stacked table definitions and datasets.  
Introduce two new components: vertical and stacked tables with a common table component.
Change the formula used to determine which format to use and now check to see if the number of stacks is larger than 8.
Stacks are created for each table cell with the same tableColumn number.
Add a HR between each table on the Summary Discharge page ... hoping this is OK with UI.  I figure this page is yet to be finalized and it helps to have the line while testing this page.
Page defs and dependant UI elements
Refactor page definitions to improve future understanding of the code.  
Let check boxes control other UI elements