MAR
Change way we get the med schedule. Add a special property into the INPUTs and use that to collect the medSchedule during definition generation. Then it's a simple matter to get this array from the med page when needed during runtime.
MAR property store updated table into page data
The Mar Summary had an extra cell in the table header due to a coding error.  
Add the vertical bar between the days in the table header