Grid work
This is a major refactoring of how inside pages manage content. It is a radical simplification over the previous methods.  Both methods are active in this code at the same time. Only the TestPage is ready for UX review. All the other pages continue to use the original methods.
In this approach a page has groups of inputs elements. These will be laid out using CSS grid. Subgroup elements are allowed.