Table transpose, seed editor view, instructor
When the instructor views student data and the "develop content" flag was on the instructor would be able to edit the student's work.  Now, the action of viewing student data disables the "develop content" flag.
The seed editor now sees a developer context banner in the EHR.
Transpose tables now work. NOTE: the way to indicate a table is transposed is via the INPUTs sheet. Add 'transpose' to the 'formOptions' for the 'table_form'.  See Current Visit -1 for an example.