Organizing folder structure
Starting to play with the organization of the documentation by user.

Bryan, I need your help with some if, then statements on the vertical navigation. I’d like different vertical navs to show up based on what the user clicks in the horizontal nav. Is that something that is possible?