Instructor dashboard -active course activity loads
When an instructor comes to the app from Moodle the course activity they want is loaded with the class list  (and not just the first assignment) and (attempt) scroll the activity into view.
Same when the instructor clicks the assignment link in the activity they are taken to the assignment page. When the user returns to the dashboard (class list) the assignment/activity that was clicked is loaded.