Selects now have keys and values
Provide means to define a key for select options. Will use this to trigger medication scheduling visibility.
In the INPUTs sheet when defining the options for a select we can now provide key := value pairs (optional!) with the special characters of := as the separator.
sched := Scheduled
prn := PRN (as needed)
once := Once
stat := Stat
document the key value definition