Activity can be closed
The instructor can now press a button on the class list page and close (or reopen) an activity for students. When the activity is closed the instructor can now evaluate all the student's work and not just the students who submitted.
The Class list page has also been reworked to show the instructor more information about the current state of the activity.
API change, adding in the closed property to the activity record.
Adding some data and time formatting directives into the system. These are used on the class list page.
A couple of TODO are added into the code.  
1. clean up the load class list methods to not use the filter property. 
2. See if the Activity component is still in use.  
3. See if we can remove the watch() in ClassList.vue
Student experience when activity is closed
- Disable edit.
- Don't show nav panel actions
- Show student evaluation panel with date closed