Implement jwt token and admin login (#651)
Implement a JWT token system to authenticate users from the time the enter via the LMS until their session expires.
Includes starting with a "refresh token" that is provided on the incoming URL. 
This token must be validated within a minute and the client will get a new token containing their visit id.
This token is needed on non public API calls.

Also provide an Admin login screen and dashboard and API end points. The admin password is 
generated and stored on the server. Only persons authorizated to access the server can obtain this password.

Admin functions include deleting records and working with tool consumers.