Fix role change issue
Problem: After user logs on as, say, a student and then returns to the LMS and comes back as an instructor the ehr system thinks the user is still a student. This happened because the vuex getter for the token cached the first accessed value (from local storage) and then did not query again. Fix is to not use Vuex getter for this property.