stage 1 demo complete
user, if they know the secret, can see the demo mode and to everything except run the LTI log in to actually try the EdEHR. They can 'log in' and out and go see a read only view of the EHR pages. They can select an assignment and a character.

Refactor icon import into a separate file

Disable demo by default