Files
UI - add file upload components with validation and styling
Validates size based on the max size as given by the server
Validates file types to be images and pdfs
Designed for a single file upload, sufficient for the needs of EdEHR
Cancel button clears
Upload button only enabled when the file is ready to send
Thumbnails of images to be uploaded are shown
File size is reported in human readable format
Last date updates is in human readable format

Change server to return JSON instead of rendered HTML.
Add error check for tool consumer id into the files api
Split UI into two views: one for adding a file and another for listing. Users may not be interested in listing and just want to add a file and return to their work.

The  demo mode does not have a full auth token with consumer id. Stuck until this is resolved.
Add unit test to file controller for the case the auth token does not have consumer
Bump file size to 5MB
Show successful upload messages
Add is authorized check. 

Change demo controller id to 'demo-timestamp-randomchars'

This helps with the file storage to identify which directories are for demo and to include a timestamp to make clean up easier. Ditto for the data in the database.
File listing error reporting on client.
Refactor demo helper to more appropriate demo store helper

To do
- [ ] add drag and drop zone (optional)
- [ ] in auth check also check user is an instructor
- [ ] Clearing out demo data, include file list
- [ ] Fix demo mode for file management

File upload show meta data related to copyright prior to send to server.
List files and view file via public API url
File list instructions and intro