Seed Editing; fix upload, allow delete, add properties: contributors and more
Fix problem with uploading a file during create and editing a seed
Clean up server side errors in prep for delete of seed
Allow deletion of a seed if there are no assignments using it.
Fix various error handlers.  Someday need to go through the app and resolve the tech debt here.