Page loading issues resolved
Loading and reloading while editing the seed was broken and page refreshes during the edit also had problems.  Divided the isDevelopingContent into two flags. One to say the user can edit course content and a new isEditingSeed to say when the user is editing a seed.  
Initialize the visit store to properly set properties that need to persist across a browser refresh.
Add query parameters to all incoming and mode changing routes.
Capture the query parameters in the common, shared page-controller loading. Manage simple state machine there.
To the router add a function to test if a route (page) is inside the EHR or not. Use to help with state machine.

Update tests