Create a EHR File Link component
Create a component to display a EhrFile whether that's in the file listing, the upload dialog, the file select dialog or in the EHR.
Have option to make this a clickable link.
Later, add ability to retrieve a thumbnail and link to full image.
Later, have a hover or other action to show a larger image without needing to open in a new tab.