Transform old admissionDay day to new visitDay format
This is the first time we've needed to make a transformation. I've done this in the client side as we refresh the page. This way the data will appear correct to anyone viewing the EHR even older content was used.
Putting the code in this location is a kludge but its good enough for now and until more transformations are needed.