Rework context banner
In preparation of providing simulation controls (care giver name, profession, encounter day and time).
These changes place the context banner under the second top level menu. This makes sense. There is the application level (Home, About, etc) then, for instructors the LMS level, and then for users viewing and working with EHR pages there is now the simulation context banner.  (May refactor the class names from EhrContext... to EhrSimulation....)

Substantial simplification of the evaluating instructor and student's experience with this area.  Cut out a lot of complicated code that attempted to do way too much. For example instructors no longer have a save and next button while evaluating students. It was too complex. And the 'submit' button on the nav panel is gone. Replaced with a submit but on the simulation level area. 
Reintroduced the collection of feedback from students after submission.