Issue page not found after submit
Some students report they see a page not found error after they submit. This patch adds the lms name and url into the submit dialog and the exit to lms action's tool tip.  We need to see these to determine the source of the PNF error.

Not that this error is only seen sometimes.