Fix loading after auth and the Courses page
After auth we must clear any previous visit id or else we reuse that. Not what is intended.

If the visit is not cleared then new activities use the assignment from the previous visit and they are not sent to the unlinked page.

If the new activity has no assignment then the Courses page throws errors on the lack of assignment and case study.