Fix issues with unlinked activities
Found that it is possible for activities to be unlinked to learning objects. This was not expected so there were many places in the code that assumed every Activity had a link to a Learning Object (called assignment in the code).  This PR addresses this situation and allows instructor users make connections for unlinked activities.