Deployment, put VOL_STORE in local .env.site; Do env check before run
Run a script that checks the environment before running docker compose.
Place the VOLUME_STORAGE env var in the local .env.site file because this is site specific, and because it makes it easy to change on the fly as each site may need to be adjusted.

Note how the backup script uses this env var.  See edehr.org/scripts_edehr/edehr_db_backup.sh)