CBC Analysis table - remove rec header
This is temporary!
The problem is both the tables on this page have rec hdrs so they both have name, profession etc. this duplication causes problems. sometimes the name/prof from the cbc appears in the pbf table as saved (and yet does not appear in the pbf dialog)
Solution is to make the element keys unique