Hide v1 tabs. Pages with tabs now display intro text. Use this for v1 tables
In the INPUTs use the tableLabel field to define a text string that will be displayed above a table's ADD button, if there is content in the field.

Use this to display a message above v1 meds, mars, lab reqs and lab reports. Message says to contact EdEHR to update the case study.