LTI. Role. Remove unused roles because they cause problems.
The Role class had an isAdmin and isDeveloper role. Neither are used but both were used in the logic to validate a request. 
If a user is the Admin of a Moodle instance and they change their role to student then the LTI request resulted in an invalid role. (Moodle sends a message without a student or instructor role but with a admin role and that caused problems.