Fix EHR file elements
The code was using an incorrect approach to sharing data between parent and child. Resolved to correctly let parent own the data, pass it to child via a prop, let child emit event when data needs to change, let parent catch event and update it's data which, with reactivity, propagates the change back into the child.