Nav system
LIS nav system stays open for all pages because it is smaller.
Fixed the size of the outside side nav when collapsed to show all content.
Adjust the wording on some tool tips.
Outside pages have better page naming. It says what the type is and then the name of the object.