push notifications
Course, Activity, ActivityData, LearningObject and CaseStudy send push notifications when there is a change via the new websocket connection. On the client side we capture several of the key events and reload data so the user sees the most up to date information.
For example, if the evaluation data is to be hidden from students (or shown) then this happens when the instructor makes the change.
When a group of instructors are grading a class the information in the class list stays the same.
When a user is working in the EHR on a table they see any changes made to that table in another window.  Unfortunately, at this time, these updates do not work for page form data (e.g. demographics)
TODO --- update case studies