v2.3.6
Introduce feedback viewable/skills assessment

The Activity's description is really a student's instructions. Change the focus of this description field into student instructions. Update the full demo activities. Show the instructions to the student when they are in the EHR

App dialog for smaller devices.
Checkset no longer errors if there is no value.
Support redirects into LIS and not just EHR.
Updates to support tabbing when it comes, Urinalysis.

Introduce websocket connection to let server push message to client
Course, Activity, ActivityData, LearningObject and CaseStudy send push notifications when there is a change via the new websocket connection. On the client side we capture several of the key events and reload data so the user sees the most up to date information.
For example, if the evaluation data is to be hidden from students (or shown) then this happens when the instructor makes the change.
When a group of instructors are grading a class the information in the class list stays the same.
When a user is working in the EHR on a table they see any changes made to that table in another window.  Unfortunately, at this time, these updates do not work for page form data (e.g. demographics)
Prevent spinner when updating activity after push msg
Trace and resolve issue of locked spinner after push message triggers updaes
TODO --- update case studies