Improve EHR form edit cancel;  Improve UI experience
The EHR form edit cancel action was terrible. It required a full page reload.  Turns out we just needed to reload the data and let reactivity handle the rest.  Not a surprise! Just something that I didn't have time to look into before.

STOP THE SPINNER.  The other annoying UI element was the darned full page - big flash -- progress spinner. Again after a little experimentation it turns out this is not needed. And not the whole UI flows much better.