Rebuild websocket client without all the event lstnrs
Previous design created an event listener per connection and that's a problem because the system has limited event listeners.
So, redesigned to use a single listener for all connections. Decided to make the client code into a class rather than a function to improve separation of concerns.

Introduce new event when user "logs in" and "logs out".  Decided to use the moment the auth header is set/unset to fire this event.
The event is useful to close websocket connections.