Significant rewrite of ehr data saving process
Saves just the user's data and doesn't mess around with any seed data if the user is a student. No need to mark the merged data. No need to extract the unmarked data.
This rewrite is essential because with multi-patient the student may create a new patient which will not have seed data.

A majority of this submission removes code and documentation that is not going to be needed.
To get the essential sense of what is happening here checkout the changes in ehr-helper around saveData