Save just student data; set up class list for student
The previous commit fixed the problem but it saves all the form data and not just what the student has entered. We now use the student's data as the basis.

On the class list page an instructor can force submit a student's work.  Recent changes meant that selecting the student did not fully load the student's data. In particular, we missed the activity data. We need its _id to send the force submit message