Skills Assessment
Instructors can enable skills assessment on one or more activities in a course.  When enabled students can are restricted to the activities allowed.

If skills assessment is active and a student user tries to visit an EHR page for an activity that is not permitted then the student user is redirected to their dashboard.

[Tech Debt] simplify the loading and storage of the Course, on the client side. Remove change for data mismatch when storing course ID and course Object as separate items. Also this simplifies the SOP to load the course.

Add helper text to Course page to explain Skills Assessment