More demo activities, MAR simplified, Vitals, Layout adjustments
Change the Demo LMS page to be more compact and readable.
Reorder the activities bringing the newer more interesting activities to the top.
Shorten the name of two sample case studies.
Make several adjustments to the Orientation case study.
Slightly enlarge the area for the medication order information on the MAR page. It was only showing 2.5 lines. Now it shows 3.
On the Vitals. Fix the day string when the day is 0. Adjust the CVP grid lines. Enlarge the font for glucose when it's off the scales high.
Fix the cursor on all dialogs. It was the move cursor which should only apply to the banner.
Adjust width of EHR pg form label in 3 cols
Update EHR content - CAM layout, Wound.
MAR page only show day selectors if more than one day is needed
EHR content layout adjustments and fixes

Make the orientation case study more realistic

Small tweaks