Seed condensed; hide v1 if not needed; show row count
Add v1 to the v1 MAR table's title so that it can be filtered.
Show a row count to tell the user how many rows there are in the table even when it is collapse.
Copy the EhrPage code to filter v1 tables to the condensed view.

hover text on collapse button