Make the instructor's life better
Allow instructors to change which learning object an activity uses. Previously we prevented this to protect them from making changes after students started working because a change in learning object changes the content.  But the restriction was causing a hardship for instructors in learning about the way the application works and made the initial encounter complicated.

Also removed the (confusing) ability to create learning objects and case studies in the same dialog. Gee that was over doing it.  Now users (content creators) can create a learning object in two places and then change the seed whenever they wish.

Also changed it so that activities always and only get their Name and Instructors from the learning object. None of this trying to offering to use the content from the LMS because (1) that content isn't usually useful and (2) it was so confusing to users which to pick and when and (3) each semester when the instructor goes to use same links they needed to go into each and every activity and set up the text to use the LO text.

Reworked the "unlinked" pages. For both the student and instructor made these a "welcome" page. I cant believe it took me this long to realize these were welcome pages. For the instructor the process to select content is much easier and more direct. And they can change their selection later so we don't need that scary warning confirmation dialog.

Create a feature flag system to keep some new feature hidden until a school is ready to use it.

Add a "unleash activity flag". When disabled (default) schools see the old behaviour. When enabled they see the new behaviour above.

All outside pages now have a "Tell me about this page" button, at the top, which toggles the display of helper text placed where it is needed through out the page. This button is a toggle for the application which makes it easy for users to learn how to use the application without leaving to see the help in the documentation.

This suggests the need to reduce the content in the docs and to direct the user to the helper text.

New dialogs for activities to find learning objects and to edit/create learning objects.

Schema changes:
Learning object (assignment) has a new calculated field idForLTI that is intended to be useful for automatic connections between LMS activities and content. Similar to the original custom parameter we had in the beginning.In fact it is the same but now it will be optional.  More about this when I implement the code in the LTI server.

Fixed page-controller to be sure the visit record is loaded for outside pages.
Adjusted a few of the demonstration activities.
Zone button now has an override to show text.
Redesign style of outside pages with new explanation text.

Change appearance of Case Study 'tags'.  Gone is the big bubble. They looked too much like buttons.  Now the tags are comma separated and slightly bolded.

Many changes to improve small screen functionality.

Themed the Case Study with a light brand red and Learning Objects with a light brand blue.

Finish patient search now searches in the seed keyData familyName field as well as the usual name and description fields.

Fix app dialog save button focus, update text through out