Visitor counting
On the LMS page show the visit count stats.
If user is Admin and is on consumer page they see same stats. Admin user can select which consumer they want to see.
The metrics API returns all consumer counts.
Add a cron task manager to the server to set up hourly tasks. Use this to schedule the retrieval of visit counts for the metric end-point.